Code merge and refactoring
